Revert "feat(datepicker): expose validators minDate, maxDate, invalidDate (#3126)" (#3219)
This reverts commit 509bba3d486b001f51eb7eddaad0123d8378663f.
