+++ src/datepicker/datepicker-input.ts
@@ -9,14 +9,16 @@ import {
   Inject,
   Input,
   NgZone,
+  OnChanges,
   OnDestroy,
   Output,
   Renderer2,
+  SimpleChanges,
   TemplateRef,
   ViewContainerRef
 } from '@angular/core';
 import {DOCUMENT} from '@angular/common';
-import {ControlValueAccessor, NG_VALUE_ACCESSOR} from '@angular/forms';
+import {AbstractControl, ControlValueAccessor, NG_VALIDATORS, NG_VALUE_ACCESSOR, Validator} from '@angular/forms';
 
 import {ngbAutoClose} from '../util/autoclose';
 import {ngbFocusTrap} from '../util/focus-trap';
@@ -37,6 +39,12 @@ const NGB_DATEPICKER_VALUE_ACCESSOR = {
   multi: true
 };
 
+const NGB_DATEPICKER_VALIDATOR = {
+  provide: NG_VALIDATORS,
+  useExisting: forwardRef(() => NgbInputDatepicker),
+  multi: true
+};
+
 /**
  * A directive that allows to stick a datepicker popup to an input field.
  *
@@ -51,10 +59,10 @@ const NGB_DATEPICKER_VALUE_ACCESSOR = {
     '(blur)': 'onBlur()',
     '[disabled]': 'disabled'
   },
-  providers: [NGB_DATEPICKER_VALUE_ACCESSOR, NgbDatepickerService]
+  providers: [NGB_DATEPICKER_VALUE_ACCESSOR, NGB_DATEPICKER_VALIDATOR, NgbDatepickerService]
 })
-export class NgbInputDatepicker implements OnDestroy,
-    ControlValueAccessor {
+export class NgbInputDatepicker implements OnChanges,
+    OnDestroy, ControlValueAccessor, Validator {
   private _cRef: ComponentRef<NgbDatepicker> = null;
   private _disabled = false;
   private _model: NgbDate;
@@ -242,6 +250,8 @@ export class NgbInputDatepicker implements OnDestroy,
 
   private _onChange = (_: any) => {};
   private _onTouched = () => {};
+  private _validatorChange = () => {};
+
 
   constructor(
       private _parserFormatter: NgbDateParserFormatter, private _elRef: ElementRef<HTMLInputElement>,
@@ -256,7 +266,31 @@ export class NgbInputDatepicker implements OnDestroy,
 
   registerOnTouched(fn: () => any): void { this._onTouched = fn; }
 
+  registerOnValidatorChange(fn: () => void): void { this._validatorChange = fn; }
+
   setDisabledState(isDisabled: boolean): void { this.disabled = isDisabled; }
+
+  validate(c: AbstractControl): {[key: string]: any} {
+    const value = c.value;
+
+    if (value === null || value === undefined) {
+      return null;
+    }
+
+    const ngbDate = this._fromDateStruct(this._dateAdapter.fromModel(value));
+
+    if (!this._calendar.isValid(ngbDate)) {
+      return {'ngbDate': {invalid: c.value}};
+    }
+
+    if (this.minDate && ngbDate.before(NgbDate.from(this.minDate))) {
+      return {'ngbDate': {requiredBefore: this.minDate}};
+    }
+
+    if (this.maxDate && ngbDate.after(NgbDate.from(this.maxDate))) {
+      return {'ngbDate': {requiredAfter: this.maxDate}};
+    }
+  }
 
   writeValue(value) {
     this._model = this._fromDateStruct(this._dateAdapter.fromModel(value));
@@ -358,6 +392,12 @@ export class NgbInputDatepicker implements OnDestroy,
   }
 
   onBlur() { this._onTouched(); }
+
+  ngOnChanges(changes: SimpleChanges) {
+    if (changes['minDate'] || changes['maxDate']) {
+      this._validatorChange();
+    }
+  }
 
   ngOnDestroy() {
     this.close();
