+++ src/datepicker/datepicker.module.ts
@@ -7,7 +7,6 @@ import {NgbDatepickerNavigation} from './datepicker-navigation';
 import {NgbInputDatepicker} from './datepicker-input';
 import {NgbDatepickerDayView} from './datepicker-day-view';
 import {NgbDatepickerNavigationSelect} from './datepicker-navigation-select';
-import {NgbInvalidDateValidator, NgbMaxDateValidator, NgbMinDateValidator} from './datepicker-validators';
 
 export {NgbDatepicker, NgbDatepickerNavigateEvent} from './datepicker';
 export {NgbInputDatepicker} from './datepicker-input';
@@ -29,19 +28,13 @@ export {NgbDateAdapter} from './adapters/ngb-date-adapter';
 export {NgbDateNativeAdapter} from './adapters/ngb-date-native-adapter';
 export {NgbDateNativeUTCAdapter} from './adapters/ngb-date-native-utc-adapter';
 export {NgbDateParserFormatter} from './ngb-date-parser-formatter';
-export {
-  NgbInvalidDateValidator,
-  NgbMaxDateValidator,
-  NgbMinDateValidator,
-  NgbDateValidators
-} from './datepicker-validators';
 
 @NgModule({
   declarations: [
     NgbDatepicker, NgbDatepickerMonthView, NgbDatepickerNavigation, NgbDatepickerNavigationSelect, NgbDatepickerDayView,
-    NgbInputDatepicker, NgbMinDateValidator, NgbMaxDateValidator, NgbInvalidDateValidator
+    NgbInputDatepicker
   ],
-  exports: [NgbDatepicker, NgbInputDatepicker, NgbMinDateValidator, NgbMaxDateValidator, NgbInvalidDateValidator],
+  exports: [NgbDatepicker, NgbInputDatepicker],
   imports: [CommonModule, FormsModule],
   entryComponents: [NgbDatepicker]
 })
