+++ src/index.ts
@@ -16,7 +16,6 @@ import {NgbTabsetModule} from './tabset/tabset.module';
 import {NgbTimepickerModule} from './timepicker/timepicker.module';
 import {NgbTooltipModule} from './tooltip/tooltip.module';
 import {NgbTypeaheadModule} from './typeahead/typeahead.module';
-import {NgbInvalidDateValidator, NgbMaxDateValidator, NgbMinDateValidator} from './datepicker/datepicker-validators';
 
 export {
   NgbAccordionModule,
@@ -54,11 +53,7 @@ export {
   NgbDateNativeAdapter,
   NgbDateNativeUTCAdapter,
   NgbDatepicker,
-  NgbDateValidators,
-  NgbInputDatepicker,
-  NgbMinDateValidator,
-  NgbMaxDateValidator,
-  NgbInvalidDateValidator
+  NgbInputDatepicker
 } from './datepicker/datepicker.module';
 export {
   NgbDropdownModule,
